+++ general/general_functions.py
@@ -234,8 +234,10 @@ class Image:
         Return the 3D image with the dimensions T, Y and X.
         When used the other dimensions F,C,Z MUST be empty (with size = 1)
     zProjection(projection_type, zrange)
-        Return the z-projection of the image using the selected projection type over the range of z values [z_best-zrange,z_best+zrange], where z_best is the Z corresponding to best focus.
-        Possible projection types: bestZ, max, min, std, avg (or mean), median
+        Return the z-projection of the image using the selected projection type over the range of z values defined by zrange.
+        Possible projection types: max, min, std, avg (or mean), median.
+        If zrange is None, use all Z values. If zrange is an integer, use z values in [z_best-zrange,z_best+zrange],
+        where z_best is the Z corresponding to best focus. If zrange is a tuple of lenght 2 (zmin,zmax), use z values in [zmin,zmax].
     """
     def __init__(self, im_path):
         self.path = im_path
@@ -305,67 +307,85 @@ class Image:
 
     def zProjection(self, projection_type, zrange):
         """
-        Return the z-projection of the image using the selected projection type over the range of z values [z_best-zrange,z_best+zrange],
-        where z_best is the Z corresponding to best focus.
+        Return the z-projection of the image using the selected projection type over the range of z values defined by zrange.
 
         Parameters
         ----------
         projection_type: str
-            the projection type (bestZ, max, min, std, avg, mean or median)
-        zrange: int
-            the number of z sections on each side of the Z with best focus
-            to use for for the projection.
+            the projection type (max, min, std, avg, mean or median)
+        zrange: int or (int,int) or None
+            the range of z sections to use for projection.
+            If zrange is None, use all z sections.
+            If zrange is an integer, use all z sections in the interval [z_best-zrange,z_best+zrange]
+            where z_best is the Z corresponding to best focus.
+            If zrange is tuple (zmin,zmax), use all z sections in the interval [zmin,zmax].
 
         Returns
         -------
         ndarray
             a 6D array with original image size, except for Z axis which has size 1.
         """
-        logging.getLogger(__name__).info('Z-Projection: projection type=%s, zrange=%s', projection_type, zrange)
-        projected_image = np.zeros((self.sizes['F'], self.sizes['T'], self.sizes['C'], 1, self.sizes['Y'], self.sizes['X']),dtype=self.image.dtype)
+        if zrange is None:
+            logging.getLogger(__name__).info('Z-Projection: projection type=%s, zrange=%s (All Z sections)', projection_type, zrange)
+        elif isinstance(zrange, int) and zrange == 0:
+            logging.getLogger(__name__).info('Z-Projection: projection type=%s, zrange=%s (Z section with best focus)', projection_type, zrange)
+        elif isinstance(zrange, int):
+            logging.getLogger(__name__).info('Z-Projection: projection type=%s, zrange=%s (Range %s around Z section with best focus)', projection_type, zrange, zrange)
+        elif isinstance(zrange, tuple) and len(zrange) == 2 and zrange[0] <= zrange[1]:
+            logging.getLogger(__name__).info('Z-Projection: projection type=%s, zrange=%s (Fixed range from %s to %s)', projection_type, zrange, zrange[0], zrange[1])
+        else:
+            logging.getLogger(__name__).info('Z-Projection: invalid zrange')
+        projected_image = np.zeros((self.sizes['F'], self.sizes['T'], self.sizes['C'], 1, self.sizes['Y'], self.sizes['X']), dtype=self.image.dtype)
         sharpness = np.zeros(self.sizes['Z'])
-        def gaus(x,a,x0,sigma,b):
+        def gaus(x, a, x0, sigma, b):
             return a*np.exp(-(x-x0)**2/(2*sigma**2))+b
         for f in range(self.sizes['F']):
             for t in range(self.sizes['T']):
                 for c in range(self.sizes['C']):
-                    for z in range(self.sizes['Z']):
-                        sharpness[z] = cv.Laplacian(self.image[f,t,c,z,:,:].astype("float64"), cv.CV_64F,ksize=11).var()
+                    z_values = None
+                    if zrange is None:
+                        # use all Z
+                        z_values = list(range(self.sizes['Z']))
+                        logging.getLogger(__name__).info('Z-Projection (F: %s, T: %s, C: %s): %s over z in %s (all)', f, t, c, projection_type, z_values)
+                    elif isinstance(zrange, int):
+                        # use zrange around Z with best focus
+                        for z in range(self.sizes['Z']):
+                            sharpness[z] = cv.Laplacian(self.image[f, t, c, z, :, :].astype("float64"), cv.CV_64F, ksize=11).var()
 
-                    ##fit a gaussian and extract position of the max
-                    try:
-                        popt,pcov = curve_fit(gaus,np.arange(sharpness.shape[0]),sharpness/max(sharpness), p0=[1,5,2,0.1])
-                        z_best = round(popt[1])
-                        fit_error=False
-                    except:
-                        z_best = int((self.sizes['Z']-1)/2)
-                        fit_error=True
+                        # fit a gaussian and extract position of the max
+                        try:
+                            popt, pcov = curve_fit(gaus, np.arange(sharpness.shape[0]), sharpness/max(sharpness), p0=[1, 5, 2, 0.1])
+                            z_best = round(popt[1])
+                            fit_error = False
+                        except:
+                            z_best = int((self.sizes['Z']-1)/2)
+                            fit_error = True
 
-
-                    if projection_type == 'bestZ':
-                        z_values = [z_best]
-                    else:
-                        #if z_best is too close to min or maz 'Z' => shift best_z so as to keep (2*zrange+1) z values (z_values).
+                        # if z_best is too close to min or maz 'Z' => shift best_z so as to keep (2*zrange+1) z values (z_values).
                         z_best_tmp = min(max(z_best, zrange), self.sizes['Z']-zrange-1)
                         z_values = [z for z in range(z_best_tmp-zrange, z_best_tmp+zrange+1) if z < self.sizes['Z'] and z >= 0 ]
 
-                    if fit_error:
-                        logging.getLogger(__name__).info('Z-Projection (F: %s, T: %s, C: %s): %s over z in %s (Best z estimation failed, using default value %s)',f, t, c, projection_type, z_values, z_best)
-                    else:
-                        logging.getLogger(__name__).info('Z-Projection (F: %s, T: %s, C: %s): %s over z in %s (Best z=%s)',f, t, c, projection_type, z_values, z_best)
+                        if fit_error:
+                            logging.getLogger(__name__).info('Z-Projection (F: %s, T: %s, C: %s): %s over z in %s (Best z estimation failed, using default value %s)', f, t, c, projection_type, z_values, z_best)
+                        else:
+                            logging.getLogger(__name__).info('Z-Projection (F: %s, T: %s, C: %s): %s over z in %s (Best z=%s)', f, t, c, projection_type, z_values, z_best)
+                    elif isinstance(zrange, tuple) and len(zrange) == 2 and zrange[0] <= zrange[1]:
+                        # use fixed range
+                        z_values = [z for z in range(zrange[0], zrange[1]+1) if z < self.sizes['Z'] and z >= 0]
+                        logging.getLogger(__name__).info('Z-Projection (F: %s, T: %s, C: %s): %s over z in %s (fixed range)', f, t, c, projection_type, z_values)
 
-                    if projection_type == 'bestZ':
-                        projected_image[f,t,c,0,:,:] = self.image[f,t,c,z_best,:,:].copy()
+                    if len(z_values) == 1:
+                        projected_image[f, t, c, 0, :, :] = self.image[f, t, c, z_values[0], :, :].copy()
                     elif projection_type == 'max':
-                        projected_image[f,t,c,0,:,:] = np.max(self.image[f,t,c,z_values,:,:], axis=0)
+                        projected_image[f, t, c, 0, :, :] = np.max(self.image[f, t, c, z_values, :, :], axis=0)
                     elif projection_type == 'min':
-                        projected_image[f,t,c,0,:,:] = np.min(self.image[f,t,c,z_values,:,:], axis=0)
+                        projected_image[f, t, c, 0, :, :] = np.min(self.image[f, t, c, z_values, :, :], axis=0)
                     elif projection_type == 'std':
-                        projected_image[f,t,c,0,:,:] = np.std(self.image[f,t,c,z_values,:,:], axis=0, ddof=1)
+                        projected_image[f, t, c, 0, :, :] = np.std(self.image[f, t, c, z_values, :, :], axis=0, ddof=1)
                     elif projection_type == 'avg' or projection_type == 'mean':
-                        projected_image[f,t,c,0,:,:] = np.mean(self.image[f,t,c,z_values,:,:], axis=0)
+                        projected_image[f, t, c, 0, :, :] = np.mean(self.image[f, t, c, z_values, :, :], axis=0)
                     elif projection_type == 'median':
-                        projected_image[f,t,c,0,:,:] = np.median(self.image[f,t,c,z_values,:,:], axis=0)
+                        projected_image[f, t, c, 0, :, :] = np.median(self.image[f, t, c, z_values, :, :], axis=0)
                     else:
                         logging.getLogger(__name__).error('Projection type not recognized')
                         return