+++ modules/registration_module/registration.py
@@ -1,6 +1,6 @@
 import os
 from PyQt5.QtCore import Qt
-from PyQt5.QtWidgets import QCheckBox, QComboBox, QFileDialog, QFormLayout, QPushButton, QVBoxLayout, QWidget, QAbstractItemView, QGridLayout, QLabel, QLineEdit, QHBoxLayout, QApplication, QSpinBox
+from PyQt5.QtWidgets import QCheckBox, QComboBox, QFileDialog, QFormLayout, QPushButton, QVBoxLayout, QWidget, QAbstractItemView, QGridLayout, QLabel, QLineEdit, QHBoxLayout, QApplication, QSpinBox, QRadioButton, QGroupBox
 from PyQt5.QtGui import QCursor, QIntValidator
 from functools import partial
 import numpy as np
@@ -34,19 +34,46 @@ class Perform(gf.Page):
         self.channel_position = QLineEdit(placeholderText='eg. 0 (default) / 1 / ...')
         self.channel_position.setMinimumWidth(200)
         self.channel_position.setValidator(QIntValidator())
+        # Z-Projection type
         self.projection_type = QComboBox(self)
-        self.projection_type.addItem("bestZ")
-        self.projection_type.setItemData(0, "keep only Z section with best focus.", Qt.ToolTipRole)
+        self.projection_type.addItem("max")
         self.projection_type.addItem("min")
         self.projection_type.addItem("mean")
         self.projection_type.addItem("median")
         self.projection_type.addItem("std")
         self.projection_type.setCurrentText("std")
-        self.projection_zrange = QSpinBox()
-        self.projection_zrange.setMinimum(0)
-        self.projection_zrange.setMaximum(20)
-        self.projection_zrange.setValue(3)
-        self.projection_zrange.setToolTip('Number of Z sections on each side of the section with best focus (bestZ) to use for projection.')
+
+        # Z-Projection range
+        # all
+        self.projection_mode_all = QRadioButton("All Z sections")
+        self.projection_mode_all.setChecked(False)
+        self.projection_mode_all.setToolTip('Project all Z sections.')
+        # only bestZ
+        self.projection_mode_bestZ = QRadioButton("Z section with best focus")
+        self.projection_mode_bestZ.setChecked(False)
+        self.projection_mode_bestZ.setToolTip('Keep only Z section with best focus.')
+        # around bestZ
+        self.projection_mode_around_bestZ = QRadioButton("Range around Z section with best focus")
+        self.projection_mode_around_bestZ.setChecked(True)
+        self.projection_mode_around_bestZ.setToolTip('Project all Z sections with Z in the interval [bestZ-range,bestZ+range], where bestZ is the Z section with best focus.')
+        self.projection_mode_around_bestZ_zrange = QSpinBox()
+        self.projection_mode_around_bestZ_zrange.setMinimum(0)
+        self.projection_mode_around_bestZ_zrange.setMaximum(20)
+        self.projection_mode_around_bestZ_zrange.setValue(3)
+        # fixed range
+        self.projection_mode_fixed = QRadioButton("Fixed range")
+        self.projection_mode_fixed.setChecked(False)
+        self.projection_mode_fixed.setToolTip('Project all Z sections with Z in the interval [from,to].')
+        self.projection_mode_fixed_zmin = QSpinBox()
+        self.projection_mode_fixed_zmin.setMinimum(0)
+        self.projection_mode_fixed_zmin.setMaximum(20)
+        self.projection_mode_fixed_zmin.setValue(4)
+        self.projection_mode_fixed_zmin.valueChanged.connect(self.projection_mode_fixed_zmin_changed)
+        self.projection_mode_fixed_zmax = QSpinBox()
+        self.projection_mode_fixed_zmax.setMinimum(0)
+        self.projection_mode_fixed_zmax.setMaximum(20)
+        self.projection_mode_fixed_zmax.setValue(6)
+        self.projection_mode_fixed_zmax.valueChanged.connect(self.projection_mode_fixed_zmax_changed)
         self.coalignment_yn_A = QCheckBox("")
         self.skip_cropping_yn_A = QCheckBox("")
         self.buttonA = QPushButton("Register")
@@ -68,8 +95,37 @@ class Perform(gf.Page):
         layout3.setFormAlignment(Qt.AlignLeft)
         layout3.addRow(" - Channel name:",self.channel_name)
         layout3.addRow(" - If needed, channel position into the c-stack:",self.channel_position)
-        layout3.addRow(" - If needed, projection for the z-stack:",self.projection_type)
-        layout3.addRow(" - If needed, projection range for the z-stack:",self.projection_zrange)
+        layout3.addRow(" - If needed, projection type for the z-stack:",self.projection_type)
+        # Z-Projection range
+        widget = QWidget()
+        layout4 = QVBoxLayout()
+        layout4.addWidget(self.projection_mode_all)
+        layout4.addWidget(self.projection_mode_bestZ)
+        layout4.addWidget(self.projection_mode_around_bestZ)
+        groupbox = QGroupBox()
+        groupbox.setToolTip('Project all Z sections with Z in the interval [bestZ-range,bestZ+range], where bestZ is the Z section with best focus.')
+        groupbox.setVisible(self.projection_mode_around_bestZ.isChecked())
+        self.projection_mode_around_bestZ.toggled.connect(groupbox.setVisible)
+        layout5 = QFormLayout()
+        layout5.addRow("Range:",self.projection_mode_around_bestZ_zrange)
+        groupbox.setLayout(layout5)
+        layout4.addWidget(groupbox)
+        layout4.addWidget(self.projection_mode_fixed)
+        groupbox = QGroupBox()
+        groupbox.setToolTip('Project all Z sections with Z in the interval [from,to].')
+        groupbox.setVisible(self.projection_mode_fixed.isChecked())
+        self.projection_mode_fixed.toggled.connect(groupbox.setVisible)
+        layout5 = QHBoxLayout()
+        layout6 = QFormLayout()
+        layout6.addRow("From:",self.projection_mode_fixed_zmin)
+        layout5.addLayout(layout6)
+        layout6 = QFormLayout()
+        layout6.addRow("To:",self.projection_mode_fixed_zmax)
+        layout5.addLayout(layout6)
+        groupbox.setLayout(layout5)
+        layout4.addWidget(groupbox)
+        widget.setLayout(layout4)
+        layout3.addRow(" - If needed, projection range for the z-stack:",widget)
         layout3.addRow(" - Co-align files with the same unique identifier (eg. smp01 for smp01_BF.nd2)", self.coalignment_yn_A)
         layout3.addRow(" - Do NOT crop aligned image", self.skip_cropping_yn_A)
         layout.addLayout(layout3)
@@ -107,7 +163,14 @@ class Perform(gf.Page):
         channel_name = self.channel_name.text()
         channel_position = self.channel_position.text()
         projection_type = self.projection_type.currentText()
-        projection_zrange = self.projection_zrange.value()
+        if self.projection_mode_all.isChecked():
+            projection_zrange = None
+        elif self.projection_mode_bestZ.isChecked():
+            projection_zrange = 0
+        elif self.projection_mode_around_bestZ.isChecked():
+            projection_zrange = self.projection_mode_around_bestZ_zrange.value()
+        elif self.projection_mode_fixed.isChecked():
+            projection_zrange = (self.projection_mode_fixed_zmin.value(), self.projection_mode_fixed_zmax.value())
         coalignment = self.coalignment_yn_A.isChecked()
         skip_crop_decision = self.skip_cropping_yn_A.isChecked()
 
@@ -192,6 +255,14 @@ class Perform(gf.Page):
         # Remove the selected file from the file list
         for item in filelist.selectedItems():
             filelist.takeItem(filelist.row(item))
+
+    def projection_mode_fixed_zmin_changed(self, value):
+        if self.projection_mode_fixed_zmax.value() < value:
+            self.projection_mode_fixed_zmax.setValue(value)
+
+    def projection_mode_fixed_zmax_changed(self, value):
+        if self.projection_mode_fixed_zmin.value() > value:
+            self.projection_mode_fixed_zmin.setValue(value)
 
 
 class Align(gf.Page):